requirements.txt
mt5linux
rpyc
flask
lightgbm
numpy
pandas
scikit-learn
python-dotenv
ta
